cleaning up /gateway but splitting into sub-directories
